chore: remove quick actions in project
